backend/feedback-ia-python/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pytest
python-dotenv
httpx